requirements.txt
streamlit
groq
pypdf
pytesseract
pillow